rename ad-prefs to privacy-prefs
