+++ static/src/javascripts/projects/common/modules/commercial/ad-prefs.lib.spec.js
@@ -16,7 +16,7 @@ jest.mock('lib/cookies', () => ({
     addCookie: jest.fn(() => null),
 }));
 
-jest.mock('common/modules/analytics/send-ad-prefs', () => ({
+jest.mock('common/modules/analytics/send-privacy-prefs', () => ({
     onConsentSet: jest.fn(() => null),
 }));
